feat(payments): WIP Stripe integration

+++ saga/src/main/kotlin/me/rasztabiga/thesis/saga/domain/command/saga/OrderLifecycleSaga.kt
@@ -136,7 +136,7 @@ class OrderLifecycleSaga {
                         unitPrice = menuItem.price
                     )
                 },
-                deliveryFee = BigDecimal.valueOf(4.20)
+                deliveryFee = 4.20
             )
         )
     }
